pyproject.toml
[project]
name = "multi-class-text-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3==1.39.15",
    "pytest==8.4.1",
    "strands-agents==1.1.0",
    "numpy==2.2.6",
    "pypdf2>=3.0.0",
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "bedrock-agentcore==0.1.1",
    "bedrock-agentcore-starter-toolkit==0.1.3",
]
